Start splitting the UAVCAN node code in several files
