+++ motor-control-firmware/src/uavcan/Reboot_handler.hpp
@@ -1,0 +1,10 @@
+#ifndef REBOOT_HANDLER_HPP
+#define REBOOT_HANDLER_HPP
+
+#include <uavcan/uavcan.hpp>
+#include <cvra/Reboot.hpp>
+#include <can-bootloader/boot_arg.h>
+
+void Reboot_handler(const uavcan::ReceivedDataStructure<cvra::Reboot> &msg);
+
+#endif
