+++ motor-control-firmware/src/uavcan/uavcan_node.h
